Combined files from yacs-integration and master
