--- index.html
@@ -1,68 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-<link rel="stylesheet" href="css.css">
-
-<title id="PageTitle">Ghost Registrar</title>
-<link rel="icon" id="PageIcon">
-
-</head>
-
-<script src="title.js"></script>
-<body onload=Generator()>
-
-<script src="register.js"></script>
-<script src="yacs.js"></script>
-
-<!-- load jQuery from Google CDN -->
-<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
-
-<div id="title">
-<h1 id="TitleText"><img id="TitleImage" width="0" height="32"></h1>
-</div>
-
-<div id="crninput">
-<form>
-  <table align="center"> <tr>
-    <td style="float:left;">CRNs:</td>
-    <td><input type="text" name="crn1" size="6" id="crn1"></td>
-    <td><input type="text" name="crn2" size="6" id="crn2"></td>
-    <td><input type="text" name="crn3" size="6" id="crn3"></td>
-    <td><input type="text" name="crn4" size="6" id="crn4"></td>
-    <td><input type="text" name="crn5" size="6" id="crn5"></td>
-    <td><input type="text" name="crn6" size="6" id="crn6"></td>
-	<td><input type="button" value="Import from YACS" onclick="ImportFromYACS();"></td>
-  </tr> </table>
-  <table align="center"> <tr>
-    <td>Registration Time:</td>
-	<td><input type="date" name="regDate" id="regDate"></td>
-	<td><input type="time" name="regTime" id="regTime"></td>
-	<td><input type="button" value="Register" onclick="RegisterSubmit();" style="float:right;"></td>
-  </tr> </table>
-</form>
-</div>
-
-<div id="regdetails">
-  <h3 id="RegTitle">Registration Details:</h3>
-  <h4>Time:</h4><p id="RegTime">
-  <h4>Date:</h4><p id="RegDate">
-  <h4>Courses:</h4>
-  <p id="Course1">
-  <p id="Course2">
-  <p id="Course3">
-  <p id="Course4">
-  <p id="Course5">
-  <p id="Course6">
-</div>
-
-<div class="footer">
-	<table width=100%> <tr> <td id=Credits align="left">Brought to you by:<br><br><br><br><br></td>
-	
-	<td align="right">
-	<a href="https://github.com/ghost-registrar/ghost-registrar">GitHub</a>
-	</td>
-	</table>
-</div>
-
-</body>
-</html>
