+++ src/index.js
@@ -1,3 +1,5 @@
 require('file-loader?name=[name].[ext]!./index.html');
 
 import './ui';
+import './yacs';
+import './register';
